requirements.txt
absl-py==0.10.0
argcomplete==1.12.3
argon2-cffi==21.1.0
astroid==2.5
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
binarytree==6.3.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
clang==5.0
click==7.1.2
configobj==5.0.6
configparser==5.0.1
coverage==6.4
cycler==0.10.0
dataclasses==0.6
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.3
dm-tree==0.1.6
docker-pycreds==0.4.0
dulwich==0.20.14
entrypoints==0.3
everett==1.0.3
flatbuffers==1.12
furl==2.1.3
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.3.3
geffnet==1.0.0
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
graphviz==0.14.2
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
importlib-resources==3.3.0
iopath==0.1.9
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
isort==5.7.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kaggle==1.5.12
kiwisolver==1.2.0
lazy-object-proxy==1.5.2
llvmlite==0.36.0
Markdown==3.3.2
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
munch==2.5.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.9
netron
notebook==6.4.3
numba==0.53.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==21.0
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.2
pathlib2==2.3.6
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plot-model==0.20
plotly==4.14.1
portalocker==2.3.2
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.13.0
psutil==5.7.3
ptflops==0.6.7
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pydot-ng==2.0.0
Pygments==2.10.0
PyJWT==2.1.0
pylint==2.7.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2020.5
PyYAML==5.3.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.24.0
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.8.0
sentry-sdk==1.4.3
setproctitle==1.2.2
setuptools-scm==6.3.2
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
text-unidecode==1.3
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
toml==0.10.2
tomli==1.2.2
-f https://download.pytorch.org/whl/torch_stable.html torch==1.9.0+cu111
torchaudio==0.9.0
torchinfo==1.6.3
torchscan==0.1.1
-f https://download.pytorch.org/whl/torch_stable.html torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.54.1
traitlets==5.1.0
typed-ast==1.4.2
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.6
wandb==0.12.17
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.1
yacs==0.1.8
yaspin==2.1.0
zipp==3.3.2